Aligning with wireframe designs
Edited formatting of header to match the design in the wireframes. Also did work to bring overall style and format of landing page and search bar area in line with design. Used a temporary placeholder image for the app logo until the final logo is available. Work still remains for style to be complete.