Style updates on SearchBar page.
Dropdown list is complete and matches design style. Clear Tags button and x button are complete and function. Tags pills container appears at the correct time, when there are active tags. In progress is making the pills populate correctly. Sample pill is in place and changes label to show the current active tag in the 0 index position.