Updated style of landing page to match design
Landing page style now matches design with the exception of the image for the assistant scryer and footer. Dropdown menu looks correct and pills for tags look correct. The one remaining issue is to make the clear buttons remove the active tags which is a conflict between activeTags and filteredTags I believe. I will continue work on this.